Refactor CognitoUserContext. Fix bug where undefined user was breaking the Card component.
