+++ src/App.js
@@ -2,16 +2,16 @@ import { ThemeProvider } from "@mui/material/styles";
 import theme from "./Theme";
 
 // Components
-import { CognitoUserContext } from 'containers/CognitoContextProvider/CognitoContextProvider';
+import { CognitoUserProvider } from 'containers/CognitoUserProvider/CognitoUserProvider';
 import Home from "pages/Home/Home";
 
 function App() {
  console.log('')
   return (
       <ThemeProvider theme={theme}>
-        <CognitoUserContext>
+        <CognitoUserProvider>
           <Home />
-        </CognitoUserContext>
+        </CognitoUserProvider>
       </ThemeProvider>
   );
 }
